Cast the long multiplations for the sum variables to doubles to keep them from rolling-over to negative integers. Also remove the redundant cast-to-double for sumY. Jason
git-svn-id: svn://orca.physics.unc.edu/Orca@4190 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Widgets/Plotting/AnalysisSupport/OR1dRoi.m
@@ -197,9 +197,9 @@ NSString* OR1dRoiCurveFitChanged = @"OR1dRoiCurveFitChanged";
 		double xDummy,y;
 		[dataSource plotter:aPlot index:x x:&xDummy y:&y];
 		sumY	+= y;
-		sumXY	+= x*y;
-		sumX2Y	+= x*x*y;
-		sumValX += y*x;
+		sumXY	+= (double)x*y;
+		sumX2Y	+= (double)x*x*y;
+		sumValX += (double)y*x;
 		
 		if (y < minY) minY = y;
 		if (y > maxY) {
@@ -211,9 +211,9 @@ NSString* OR1dRoiCurveFitChanged = @"OR1dRoiCurveFitChanged";
 	
 	
 	if(totalNum){
-		double theXAverage = sumXY / (double)sumY;
+		double theXAverage = sumXY / sumY;
 		average = theXAverage;
-		sigma	= sqrt((sumX2Y/(double)sumY) - (theXAverage*theXAverage));
+		sigma	= sqrt((sumX2Y/sumY) - (theXAverage*theXAverage));
 	}
 	else {
 		average = 0;
